# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trip Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new traveler (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/{userId}/change-password:
    post:
      operationId: changeUserPassword
      tags: [users]
      summary: Change a user's password (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List trips for current user
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{tripId}:
    get:
      operationId: getTrip
      tags: [trips]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    patch:
      operationId: updateTrip
      tags: [trips]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /trips/{tripId}/summary:
    get:
      operationId: getTripSummary
      tags: [trips]
      summary: Get trip overview stats
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripSummary"
  /trips/{tripId}/timeline:
    get:
      operationId: getTripTimeline
      tags: [trips]
      summary: Get all trip events in chronological order
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
  /trips/{tripId}/participants:
    get:
      operationId: listTripParticipants
      tags: [trips]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: addTripParticipant
      tags: [trips]
      summary: Add a traveler to a trip (admin only)
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParticipantInput"
      responses:
        "200":
          description: Participant added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /trips/{tripId}/participants/{userId}:
    delete:
      operationId: removeTripParticipant
      tags: [trips]
      summary: Remove a traveler from a trip (admin only)
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Participant removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    User:
      type: object
      required: [id, username, name, role]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, traveler]
        email:
          type: ["string", "null"]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [username, name, password]
      properties:
        username:
          type: string
        name:
          type: string
        password:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, traveler]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
    ChangePasswordInput:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string
    Trip:
      type: object
      required: [id, title, destination, startDate, endDate, status]
      properties:
        id:
          type: integer
        title:
          type: string
        destination:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        description:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        status:
          type: string
          enum: [planning, confirmed, active, completed]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
    TripInput:
      type: object
      required: [title, destination, startDate, endDate]
      properties:
        title:
          type: string
        destination:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        description:
          type: string
        coverImage:
          type: string
        status:
          type: string
          enum: [planning, confirmed, active, completed]
        adminNotes:
          type: string
    TripUpdate:
      type: object
      properties:
        title:
          type: string
        destination:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        description:
          type: string
        coverImage:
          type: string
        status:
          type: string
          enum: [planning, confirmed, active, completed]
        adminNotes:
          type: string
    TripSummary:
      type: object
      required: [tripId, daysCount, flightsCount, accommodationsCount, activitiesCount, participantsCount, packingItemsCount, packingCheckedCount]
      properties:
        tripId:
          type: integer
        daysCount:
          type: integer
        flightsCount:
          type: integer
        accommodationsCount:
          type: integer
        activitiesCount:
          type: integer
        participantsCount:
          type: integer
        packingItemsCount:
          type: integer
        packingCheckedCount:
          type: integer
    TimelineEvent:
      type: object
      required: [id, type, date, title]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [flight, accommodation, activity, itinerary]
        date:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        time:
          type: ["string", "null"]
    ParticipantInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
...