# lib/api-zod/src/generated/types/timelineEvent.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Trip Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { TimelineEventType } from './timelineEventType';

export interface TimelineEvent {
  id: number;
  type: TimelineEventType;
  date: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  time?: string | null;
}